requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1.1
django-extensions==3.2.1
django-filter==22.1
django-graphos==0.3.41
djangorestframework==3.13.1
idna==3.3
install==1.3.5
psycopg2==2.9.3
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
